% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usage-tracking.R
\name{ask_usage_tracking_permission}
\alias{ask_usage_tracking_permission}
\title{Ask user for permission to enable usage tracking}
\usage{
ask_usage_tracking_permission()
}
\value{
Logical indicating user's choice
}
\description{
Ask user for permission to enable usage tracking
}
\keyword{internal}
