% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usage-tracking.R
\name{check_usage_alerts_realtime}
\alias{check_usage_alerts_realtime}
\title{Check usage alerts in real-time}
\usage{
check_usage_alerts_realtime(usage_data, current_call_type)
}
\arguments{
\item{usage_data}{Data frame with current usage data}

\item{current_call_type}{Character string of the call type just made}
}
\description{
Check usage alerts in real-time
}
\keyword{internal}
