% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usage-tracking.R
\name{get_usage_file_path}
\alias{get_usage_file_path}
\title{Get the path to the usage tracking file}
\usage{
get_usage_file_path()
}
\value{
Character string with the full path to the usage file
}
\description{
Get the path to the usage tracking file
}
\keyword{internal}
