% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usage-tracking.R
\name{init_session_usage}
\alias{init_session_usage}
\title{Initialize session usage data}
\usage{
init_session_usage()
}
\value{
Data frame with correct structure
}
\description{
Initialize session usage data
}
\keyword{internal}
