% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvk_API.R
\name{inlezen_kvk}
\alias{inlezen_kvk}
\title{Read KVK-API}
\usage{
inlezen_kvk()
}
\value{
Tibble
}
\description{
Test to make sure the environmental variables on github work to build the
pkgdown website.
}
\examples{

x <- inlezen_kvk()
print(x)
}
