% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvk_API.R
\name{kvk_set_api_key}
\alias{kvk_set_api_key}
\title{Set KvK API Key for the current session}
\usage{
kvk_set_api_key(api_name, api_key)
}
\arguments{
\item{api_name}{A string containing the name of the API. One of: "search",
"basisprofiel", "vestigingsprofiel", "naamgeving".}

\item{api_key}{A string containing the API key for the specified API.}
}
\value{
Invisibly returns \code{TRUE} if the key was set.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function sets the specified KvK API key for the \strong{current R session}
using \code{Sys.setenv()}. The key will be available until the session ends.
}
\details{
The function sets the specified API key using \code{Sys.setenv()}, making
it available for the current session. However, it does \strong{not} persist
across sessions.

If you want to store the key permanently, use \code{kvk_store_api_key()}.
}
\examples{
\dontrun{
# Set the API key for the current session
kvk_set_api_key("search", "abcd1234")
}

}
