% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvkapiR-package.R
\docType{package}
\name{kvkapiR-package}
\alias{kvkapiR}
\alias{kvkapiR-package}
\title{kvkapiR: Interface to the Dutch Chamber of Commerce (KvK) API}
\description{
Access business registration data from the Dutch Chamber of Commerce (Kamer van Koophandel, KvK) through their official API \url{https://developers.kvk.nl/}. Search for companies by name, location, or registration number. Retrieve detailed business profiles, establishment information, and company name histories. Built on 'httr2' for robust API interaction with automatic pagination, error handling, and usage tracking.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://coeneisma.github.io/kvkapiR/}
  \item \url{https://github.com/coeneisma/kvkapiR/}
  \item Report bugs at \url{https://github.com/coeneisma/kvkapiR/issues}
}

}
\author{
\strong{Maintainer}: Coen Eisma \email{coeneisma@gmail.com} (\href{https://orcid.org/0009-0007-9001-2572}{ORCID}) [copyright holder]

}
\keyword{internal}
