% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usage-tracking.R
\name{prepare_monthly_usage}
\alias{prepare_monthly_usage}
\title{Prepare monthly usage summary}
\usage{
prepare_monthly_usage(usage_data)
}
\arguments{
\item{usage_data}{Data frame with usage data}
}
\value{
Tibble with monthly summary
}
\description{
Prepare monthly usage summary
}
\keyword{internal}
