% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usage-tracking.R
\name{prepare_session_usage}
\alias{prepare_session_usage}
\title{Prepare session usage summary}
\usage{
prepare_session_usage(usage_data)
}
\arguments{
\item{usage_data}{Data frame with usage data}
}
\value{
Tibble with session summary
}
\description{
Prepare session usage summary
}
\keyword{internal}
