% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usage-tracking.R
\name{save_usage_data}
\alias{save_usage_data}
\title{Save usage data to session storage}
\usage{
save_usage_data(usage_data)
}
\arguments{
\item{usage_data}{Data frame with usage data to save}
}
\description{
Save usage data to session storage
}
\keyword{internal}
