% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usage-tracking.R
\name{usage_tracking_enabled}
\alias{usage_tracking_enabled}
\title{Check if usage tracking is enabled}
\usage{
usage_tracking_enabled()
}
\value{
Logical indicating if tracking is enabled
}
\description{
Check if usage tracking is enabled
}
\keyword{internal}
