% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usage-tracking.R
\name{display_usage_summary}
\alias{display_usage_summary}
\title{Display usage summary}
\usage{
display_usage_summary(usage_data)
}
\arguments{
\item{usage_data}{Data frame with usage data}
}
\description{
Display usage summary
}
\keyword{internal}
