% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvk_API.R
\name{kvk_search}
\alias{kvk_search}
\title{Retrieve all results from the KvK Search API (up to 1.000 results)}
\usage{
kvk_search(...)
}
\arguments{
\item{...}{Named arguments passed to the API query (e.g., naam = "Koudum").
Available arguments can be found at
\url{https://developers.kvk.nl/documentation/zoeken-api#input}.}
}
\value{
A tibble containing the retrieved results. Possible parameters can be
found under the \code{Results} section of
\url{https://developers.kvk.nl/documentation/zoeken-api#output}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function automatically paginates through the KvK API to collect
available results. Due to API limitations, it retrieves a maximum of 1.000
records. When this happens, a warning will be displayed.
}
\examples{

koudum <- kvk_search(plaats = "Koudum")
koudum

rotterdam <- kvk_search(plaats = "Rotterdam")
rotterdam
}
