% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvkapiR-package.R
\docType{package}
\name{kvkapiR-package}
\alias{kvkapiR}
\alias{kvkapiR-package}
\title{kvkapiR: An R Wrapper for the Dutch Chamber of Commerce API}
\description{
Provides a convenient interface to the Dutch Chamber of Commerce (KvK) API. This package is built using the `httr2` package and follows best practices for wrapping APIs in R, as outlined in the httr2 documentation (https://httr2.r-lib.org/articles/wrapping-apis.html). It simplifies authentication, request handling, and response parsing when interacting with the KvK API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://coeneisma.github.io/kvkapiR/}
  \item \url{https://github.com/coeneisma/kvkapiR/}
}

}
\author{
\strong{Maintainer}: Coen Eisma \email{coeneisma@gmail.com} (\href{https://orcid.org/0009-0007-9001-2572}{ORCID})

}
\keyword{internal}
